+++ tests/test_visitors/test_complexity/test_offset_visitor.py
@@ -13,6 +13,12 @@ def container():
         x = 1
 """
 
+nested_if2 = """
+def container():
+    if some_value:
+        call_other()
+"""
+
 nested_for = """
 def container():
     for i in '123':
@@ -25,6 +31,15 @@ def container():
         some()
     except Exception:
         raise
+"""
+
+nested_try2 = """
+def container():
+    if some:
+        try:
+            some()
+        except Exception:
+            raise
 """
 
 nested_with = """
@@ -43,8 +58,10 @@ while True:
 
 @pytest.mark.parametrize('code', [
     nested_if,
+    nested_if2,
     nested_for,
     nested_try,
+    nested_try2,
     nested_with,
     nested_while,
 ])
@@ -60,8 +77,10 @@ def test_nested_offset(assert_errors, parse_ast_tree, code, default_options):
 
 @pytest.mark.parametrize('code, number_of_errors', [
     (nested_if, 1),
+    (nested_if2, 1),
     (nested_for, 1),
     (nested_try, 2),
+    (nested_try2, 4),
     (nested_with, 1),
     (nested_while, 1),
 ])