+++ wemake_python_styleguide/compat.py
@@ -14,6 +14,6 @@ def maybe_set_parent(tree: ast.AST) -> ast.AST:
     for statement in ast.walk(tree):
         for child in ast.iter_child_nodes(statement):
             if not hasattr(child, 'parent'):  # noqa: Z113
-                setattr(child, 'parent', statement)  # noqa: Z113
+                setattr(child, 'parent', statement)
 
     return tree